Fixed PSG drum phasing issue + moved SCC player state to 0xee00
The PSG kick/snare drum transient could sometime occur during the 'off'' phase of the oscillator, which would cause the transient to not be there, or to lose its bite. We fix this by changing the frequency of those drums to '0' so that the oscillator is permanently in the 'on 'state. (note: for the snares this only applies to the initial 'click' part of the sound, not the noise part)
I managed to find some memory for the SCC player's state in 0xee00-0xf100. This is where Vampire Killer normally stores the stack, so I moved the stack to 0xfaf0 which is an unused area of memory (normally used by MSX BASIC). This way we no longer need to change page 0 to RAM, which allows us to remove a lot of complicated slot manipulation code and potential compatibility issues.